Test commit: SVG path broken on VS Code
