+++ npm/qsharp/ux/scatterChart.tsx
@@ -2,7 +2,9 @@
 // Licensed under the MIT License.
 
 import { useRef, useEffect } from "preact/hooks";
+import { createRef } from "preact";
 import * as utils from "../src/utils.js";
+import { saveToPng, saveToSvg } from "./saveImage.js";
 
 export type ScatterSeries = {
   color: string;
@@ -105,7 +107,7 @@ export function ScatterChart(props: {
     const centerX = (pointRect.left + pointRect.right) / 2;
     const divRect = topDiv.getBoundingClientRect();
     tooltip.style.left = `${centerX - divRect.left - halfWidth}px`;
-    tooltip.style.top = `${centerY - divRect.top + 12}px`;
+    tooltip.style.top = `${centerY - divRect.top}px`;// + 12}px`;
     tooltip.style.visibility = "visible";
   }
 
@@ -154,6 +156,46 @@ export function ScatterChart(props: {
   }
   const selectedPoint = getSelectedPointData();
 
+  const saveRef = createRef();
+
+  const handleSaveImage = async () => {
+    const element = saveRef.current;
+    const backgroundColor =
+      getComputedStyle(element).getPropertyValue("--main-background");
+    const checkVSCodeEnv = getComputedStyle(element).getPropertyValue(
+      "--vscode-foreground",
+    );
+    const isVSCodeEnv = true ? checkVSCodeEnv : false;
+
+    if (isVSCodeEnv) {
+      const data = await saveToPng(element, backgroundColor);
+      const link = document.createElement("a");
+      if (typeof link.download === "string") {
+        link.href = data;
+        link.download = "image.png";
+
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+      } else {
+        window.open(data);
+      }
+    } else {
+      const data = await saveToSvg(element);
+      const link = document.createElement("a");
+      if (typeof link.download === "string") {
+        link.href = data;
+        link.download = "image.svg";
+
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+      } else {
+        window.open(data);
+      }
+    }
+  };
+
   // Need to render first to get the element layout to position the tooltip
   useEffect(() => {
     if (!selectedTooltipDiv.current) return;
@@ -182,6 +224,7 @@ export function ScatterChart(props: {
         onMouseOver={(ev) => onPointMouseEvent(ev, "over")}
         onMouseOut={(ev) => onPointMouseEvent(ev, "out")}
         onClick={(ev) => onPointMouseEvent(ev, "click")}
+        ref={saveRef}
       >
         <line
           class="qs-scatterChart-axis"
@@ -316,6 +359,25 @@ export function ScatterChart(props: {
       </svg>
       <div class="qs-scatterChart-selectedInfo" ref={selectedTooltipDiv}></div>
       <div class="qs-scatterChart-tooltip"></div>
+      <button
+        role="button"
+        onClick={handleSaveImage}
+        className={"qs-estimatesOverview-saveIcon"}
+      >
+        <span>
+          <svg
+            width="75%"
+            height="75%"
+            viewBox="0 0 16 16"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              className={"qs-estimatesOverview-saveIconSvgPath"}
+              d="M12.0147 2.8595L13.1397 3.9845L13.25 4.25V12.875L12.875 13.25H3.125L2.75 12.875V3.125L3.125 2.75H11.75L12.0147 2.8595ZM3.5 3.5V12.5H12.5V4.406L11.5947 3.5H10.25V6.5H5V3.5H3.5ZM8 3.5V5.75H9.5V3.5H8Z"
+            />
+          </svg>
+        </span>
+      </button>
     </div>
   );
 }
